app/api/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
scikit-learn==1.8.0
numpy<2.0
joblib==1.3.1
reportlab==4.0.4
groq==0.4.1
python-dotenv==1.0.0
